requirements.txt
absl-py==0.9.0
amqp==2.5.2
appdirs==1.4.3
astor==0.8.1
astunparse==1.6.3
atomicwrites==1.4.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.0
billiard==3.6.3.0
bleach==3.3.0
bs4==0.0.1
cachetools==4.0.0
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
client==0.0.1
coc==0.2
colorama==0.4.3
configparser==5.0.0
contextlib2==0.6.0.post1
cycler==0.10.0
Cython==0.29.17
dask==2.16.0
data==0.4
dataclasses==0.6
decorator==4.4.1
decouple==0.0.7
defusedxml==0.6.0
dill==0.3.1.1
distlib==0.3.0
dj-database-url==0.5.0
Django==2.2.13
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-mathfilters==1.0.0
dm-tree==0.1.5
docker-pycreds==0.4.0
entrypoints==0.3
filelock==3.0.12
funcsigs==1.0.2
future==0.18.2
gast==0.3.3
gin-config==0.3.0
gitdb==4.0.5
GitPython==3.1.2
google==2.0.3
google-api-core==1.17.0
google-api-python-client==1.8.3
google-auth==1.11.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.24.0
google-cloud-core==1.3.0
google-pasta==0.2.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
gql==0.2.0
graphql-core==1.1
graphviz==0.13.2
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
httplib2==0.18.0
idna==2.9
imageio==2.8.0
importlib-metadata==1.5.0
ipydatawidgets==4.0.1
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==1.2.6
jupyterlab-server==1.0.6
K3D==2.8.1
kaggle==1.5.6
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-tuner==1.0.1
kiwisolver==1.1.0
kombu==4.6.8
llvmlite==0.31.0
lxml==4.6.2
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
mlfq-sim==0.1.1.post0.dev58+gb5916aa
mlperf-compliance==0.0.10
more-itertools==8.3.0
mysql-connector-python==2.0.4
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
nibabel==3.1.0
nltk==3.4.5
notebook==6.1.5
numba==0.48.0
numpy==1.18.1
nvidia-ml-py3==7.352.0
oauth2client==4.1.3
oauthlib==3.1.0
object-detection @ file:///C:/Users/mitts/Desktop/object/models/research
opencv-python==4.2.0.34
opencv-python-headless==4.2.0.34
opt-einsum==3.2.0
packaging==20.4
panda==0.3.1
pandas==1.0.1
pandas-datareader==0.8.1
pandocfilters==1.4.2
parso==0.6.1
pathtools==0.1.2
patsy==0.5.1
pickleshare==0.7.5
Pillow==5.1.0
pipenv==2018.11.26
pluggy==0.13.1
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==3.0.3
protobuf==3.11.3
psutil==5.7.0
psycopg2==2.8.5
py==1.8.1
py-cpuinfo==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydicom==1.4.2
Pygments==2.5.2
pyparsing==2.4.6
PyQt5==5.14.2
PyQt5-sip==12.7.2
pyrsistent==0.15.7
pytest==5.4.2
python-dateutil==2.8.1
python-slugify==4.0.0
pytz==2019.3
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3
pyzmq==18.1.1
qtconsole==4.7.4
QtPy==1.9.0
queuelib==1.5.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
sentencepiece==0.1.90
sentry-sdk==0.14.4
shortuuid==1.0.1
six==1.14.0
sklearn==0.0
smmap==3.0.4
soupsieve==2.0
sqlparse==0.3.1
statsmodels==0.11.0
subprocess32==3.5.4
suds==0.4
tabulate==0.8.7
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.1
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.1
tensorflow-gpu-estimator==2.2.0
tensorflow-hub==0.8.0
tensorflow-metadata==0.22.0
tensorflow-model-optimization==0.3.0
termcolor==1.1.0
terminado==0.8.3
terminaltables==3.1.0
termine==0.1.5
testpath==0.4.4
text-unidecode==1.3
tf-models-official==2.2.0
tornado==6.0.3
tqdm==4.45.0
traitlets==4.3.3
traittypes==0.2.1
typeguard==2.7.1
typing==3.7.4.1
unicorn==1.0.1
uritemplate==3.0.1
urllib3==1.24.3
vine==1.3.0
virtualenv==20.0.17
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
wandb==0.8.36
watchdog==0.10.2
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
whitenoise==5.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.0.0
